wip: refactor firebase implementation
